+++ KSP2Runtime/KSPVessel/KSPVesselModule.Autopilot.cs
@@ -38,17 +38,17 @@ namespace KontrolSystem.KSP.Runtime.KSPVessel {
             public Vector3d TargetOrientation {
                 get {
                     var sas = vesselAdapter.vessel.Autopilot?.SAS;
-                    return sas != null ? vesselAdapter.vessel.transform.GetSimSOIBodyParentTransformFrame()
+                    return sas != null ? vesselAdapter.vessel.mainBody.coordinateSystem
                         .ToLocalVector(sas.ReferenceFrame, sas.TargetOrientation) : vesselAdapter.Facing.Vector;
                 }
-                set => vesselAdapter.vessel.Autopilot?.SAS?.SetTargetOrientation(new Vector(vesselAdapter.vessel.transform.GetSimSOIBodyParentTransformFrame(), value), false);
+                set => vesselAdapter.vessel.Autopilot?.SAS?.SetTargetOrientation(new Vector(vesselAdapter.vessel.mainBody.coordinateSystem, value), false);
             }
 
             [KSField]
             public Direction LockDirection {
                 get {
                     var sas = vesselAdapter.vessel.Autopilot?.SAS;
-                    return sas != null ? new Direction(vesselAdapter.vessel.transform.GetSimSOIBodyParentTransformFrame()
+                    return sas != null ? new Direction(vesselAdapter.vessel.mainBody.coordinateSystem
                         .ToLocalRotation(sas.ReferenceFrame, sas.LockedRotation)) : vesselAdapter.Facing;
                 }
                 set => vesselAdapter.vessel.Autopilot?.SAS?.LockRotation(value.Rotation);